requirements.txt
fastapi
uvicorn
supabase==2.10.0
python-dotenv
requests
pydantic
